Renamed "richEditorStyles" to "richEditorClasses" IE 11 compatibility
