+++ plugins/rich-editor/src/scripts/flyouts/EmojiFlyout.tsx
@@ -10,7 +10,7 @@ import FlyoutToggle, { IFlyoutToggleChildParameters } from "@library/flyouts/Fly
 import { t } from "@library/utility/appUtils";
 import classNames from "classnames";
 import EmojiPicker from "@rich-editor/flyouts/pieces/EmojiPicker";
-import { richEditorClasses } from "@rich-editor/editor/richEditorClasses";
+import { richEditorClasses } from "@rich-editor/editor/richEditorStyles";
 import { forceSelectionUpdate } from "@rich-editor/quill/utility";
 import { IconForButtonWrap } from "@rich-editor/editor/pieces/IconForButtonWrap";
 import { ButtonTypes } from "@library/forms/buttonStyles";
