+++ plugins/rich-editor/src/scripts/flyouts/pieces/EmojiPicker.tsx
@@ -17,7 +17,7 @@ import { insertEmojiClasses } from "@rich-editor/flyouts/pieces/insertEmojiClass
 import { EmojiGroupButton } from "@rich-editor/flyouts/pieces/EmojiGroupButton";
 import EmojiButton from "@rich-editor/flyouts/pieces/EmojiButton";
 import Flyout from "@rich-editor/flyouts/pieces/Flyout";
-import { richEditorClasses } from "@rich-editor/editor/richEditorClasses";
+import { richEditorClasses } from "@rich-editor/editor/richEditorStyles";
 import { richEditorVariables } from "@rich-editor/editor/richEditorVariables";
 
 const BUTTON_SIZE = richEditorVariables().sizing.emojiSize;
