+++ plugins/rich-editor/src/scripts/menuBar/paragraph/tabs/ParagraphMenuBarTab.tsx
@@ -7,7 +7,7 @@
 import React from "react";
 import classNames from "classnames";
 import ScreenReaderContent from "@library/layout/ScreenReaderContent";
-import { richEditorClasses } from "@rich-editor/editor/richEditorClasses";
+import { richEditorClasses } from "@rich-editor/editor/richEditorStyles";
 import { IconForButtonWrap } from "@rich-editor/editor/pieces/IconForButtonWrap";
 
 interface IProps {
