+++ plugins/rich-editor/src/scripts/menuBar/paragraph/tabs/ParagraphMenuSpecialBlockTabContent.tsx
@@ -10,7 +10,7 @@ import ParagraphMenuBarRadioGroup, {
     IMenuBarRadioButton,
 } from "@rich-editor/menuBar/paragraph/items/ParagraphMenuBarRadioGroup";
 import classNames from "classnames";
-import { richEditorClasses } from "@rich-editor/editor/richEditorClasses";
+import { richEditorClasses } from "@rich-editor/editor/richEditorStyles";
 
 interface IProps {
     items: IMenuBarRadioButton[];
