+++ plugins/rich-editor/src/scripts/toolbars/pieces/MentionSuggestion.tsx
@@ -8,7 +8,7 @@ import React from "react";
 import classNames from "classnames";
 import { t } from "@library/utility/appUtils";
 import { splitStringLoosely } from "@vanilla/utils";
-import { richEditorClasses } from "@rich-editor/editor/richEditorClasses";
+import { richEditorClasses } from "@rich-editor/editor/richEditorStyles";
 import { IUserFragment } from "@library/@types/api/users";
 
 export interface IMentionSuggestionData extends IUserFragment {
