+++ plugins/rich-editor/src/scripts/toolbars/pieces/MenuItem.tsx
@@ -6,7 +6,7 @@
 
 import React from "react";
 import classnames from "classnames";
-import { richEditorClasses } from "@rich-editor/editor/richEditorClasses";
+import { richEditorClasses } from "@rich-editor/editor/richEditorStyles";
 import { IconForButtonWrap } from "@rich-editor/editor/pieces/IconForButtonWrap";
 
 export interface IMenuItemData {
